\name{getRemoteURL}
\alias{getRemoteURL}
\alias{getLocalDir}

%- Also NEED an '\alias' for EACH other topic documented here.
\title{Setting Remote URLs}
\description{
  Tools for getting/setting remote URLs for remote databases
}
\usage{
getRemoteURL()
getLocalDir()
}
%- maybe also 'usage' for other objects documented here.
\value{
  \code{getRemoteURL()} returns the URL for the remote collection of
  databases.  \code{getLocalDir()} returns the directory on the local
  machine where copies of remote databases will be stored.
}
\author{Roger D. Peng \email{rpeng@jhsph.edu}}


% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{utilities}
